chore: refactor icons generator, fix ESLint errors
